requirements.txt
pyautogui
Pillow
opencv-python
numpy
flask
imutils
mss
sockets
eel
keyboard
obs-websocket-py
progress

#below only used for developement
#htmlmin
#jsmin